+++ R/globals.R
@@ -1,4 +1,4 @@
-if (getRversion() >= "2.15.1") utils::globalVariables(c("P.value", "PValue", "Term", "gene", "label", "logFC", "metadata"))
+if (getRversion() >= "2.15.1") utils::globalVariables(c("P.value", "PValue", "Term", "Gene", "gene", "label", "logFC", "metadata", ".group", "Expression"))
 
 #' @importFrom stats na.omit reorder sd setNames
 #' @importFrom utils head read.delim