+++ R/zzz.R
@@ -5,5 +5,3 @@
     "Documentation: https://francesco-e-vallone.github.io/TranscripTools/"
   )
 }
-
-utils::globalVariables(c("Gene", ".group", "Expression"))